Lint headers (#3)
* You should lint the headers.

* Linted headers. My bad.

* Trying later travis build.

* FUCK IT.

* Don't run linter if uncrustify version is too crusty.

* Revert "FUCK IT."

* Print uncrustify version.

+++ test/lint/run-linter.bash
@@ -15,7 +15,7 @@ fi
 
 # make sure we are in the right place
 if [[ ! -f LICENSE ]]; then
-    echo - "\e[31mRun linter script from the root of the repo.\e[0m"
+    echo -e "\e[31mRun linter script from the root of the repo.\e[0m"
     exit 2
 fi
 
@@ -26,6 +26,39 @@ for c_file in src/*.c; do
 
     if [[ ! -z $dif ]]; then
         echo -e "\[31m$c_file does not conform to Kernel style!"
+        echo "Patch will be found in $c_file.patch."
+        diff -ur $c_file $c_file.uncrustify > $c_file.patch
+        rm $c_file.uncrustify
+        echo "You may apply the patch like this:"
+        echo -e "    patch -p1 $c_file < $c_file.patch\e[0m"
+        exit 3
+    fi
+    rm $c_file.uncrustify
+done
+
+# version check
+# TODO(allenh1): when travis leaves trusty, switch to this.
+# temp_file=$(mktemp)
+# uncrustify --version > ${temp_file}
+# sed -i -e 's/Uncrustify-//g' ${temp_file}
+# sed -i -e 's/_[a-zA-Z]//g' ${temp_file}
+# version=$(cat ${temp_file})
+# rm -f ${temp_file}
+ver_string=$(uncrustify --version | awk '{print $1}')
+
+if [[ $ver_string = *'uncrustify'* ]]; then
+    echo -e "\e[31mUncrustify cannot test headers!\e[0m"
+    echo -e "\e[32mPretending things are fine...\e[0m"
+    exit 0
+fi
+
+# Lint the headers
+for c_file in include/*/*.h; do
+    uncrustify -c $config $c_file
+    dif=$(diff -ur $c_file $c_file.uncrustify)
+
+    if [[ ! -z $dif ]]; then
+        echo -e "\e[31m$c_file does not conform to Kernel style!"
         echo "Patch will be found in $c_file.patch."
         diff -ur $c_file $c_file.uncrustify > $c_file.patch
         rm $c_file.uncrustify
